+++ app/controllers/claimToAdjustPoa/EnterPoAAmountController.scala
@@ -27,7 +27,8 @@ import play.api.Logger
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents, Result}
 import services.{ClaimToAdjustService, PaymentOnAccountSessionService}
 import uk.gov.hmrc.auth.core.AuthorisedFunctions
-import utils.{AuthenticatorPredicate, ClaimToAdjustUtils, JourneyCheckerClaimToAdjust}
+import utils.ClaimToAdjust.{ClaimToAdjustUtils, JourneyCheckerClaimToAdjust}
+import utils.AuthenticatorPredicate
 import views.html.claimToAdjustPoa.EnterPoAAmountView
 
 import javax.inject.{Inject, Singleton}
