+++ app/controllers/claimToAdjustPoa/WhatYouNeedToKnowController.scala
@@ -26,8 +26,8 @@ import play.api.i18n.I18nSupport
 import play.api.mvc._
 import services.{ClaimToAdjustService, PaymentOnAccountSessionService}
 import uk.gov.hmrc.auth.core.AuthorisedFunctions
-import utils.ClaimToAdjust.WithSessionAndPoa
-import utils.{AuthenticatorPredicate, ClaimToAdjustUtils}
+import utils.ClaimToAdjust.{ClaimToAdjustUtils, WithSessionAndPoa}
+import utils.AuthenticatorPredicate
 import views.html.claimToAdjustPoa.WhatYouNeedToKnow
 
 import javax.inject.{Inject, Singleton}
